Mid-semester exam program (Arrays)
For my mid-semester exam in C, I was asked to write a C program that reads the oral and written exam grades of 10 students (from the keyboard), and stores them in a 10x2 array. Then the average of each student is calculated and stored in a simple array. Lastly, the maximum average grade is printed.